+++ java/basic/Distance.java
@@ -7,7 +7,7 @@ public static void main(String[]args) {
     double distance = Math.sqrt((Math.pow((x2-x1), 2) + Math.pow((y2-y1),2)));
 
 
-    System.out.println("distance of 2 point :"+ distance);
+    System.out.println("distance of 2 point :"+ (int)distance);
 }
     
 }
